osint/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
requests==2.32.3
pyyaml==6.0.2
